Unify structure of the monad code * Lazy got its own directory too. * Moved ExceptionUtilities to Compatibility. * Extracted the LINQ special Monad Functions (Select and SelectMany) into their own files. * Extracted the Monad Law Tests into their own files.
