+++ Funcky.Test/Monads/OptionTest.cs
@@ -322,44 +322,6 @@ namespace Funcky.Test.Monads
         }
 
         [Theory]
-        [MemberData(nameof(OptionValues))]
-        public void AssociativityHolds(Option<int> input)
-        {
-            static Option<int> MultiplyByTen(int x) => Option.Some(x * 10);
-            static Option<int> AddTwo(int x) => Option.Some(x + 2);
-            static Option<int> MultiplyByTenAndAddTwo(int x) => MultiplyByTen(x).SelectMany(AddTwo);
-
-            Assert.Equal(
-                input.SelectMany(MultiplyByTen).SelectMany(AddTwo),
-                input.SelectMany(MultiplyByTenAndAddTwo));
-        }
-
-        [Theory]
-        [MemberData(nameof(OptionValues))]
-        public void RightIdentityHolds(Option<int> input)
-        {
-            Assert.Equal(input.SelectMany(Option.Some), input);
-        }
-
-        public static TheoryData<Option<int>> OptionValues()
-            => new()
-            {
-                Option.Some(10),
-                Option<int>.None(),
-            };
-
-        [Fact]
-        public void LeftIdentityHolds()
-        {
-            static Option<int> MultiplyByTen(int x) => Option.Some(x * 10);
-
-            const int input = 10;
-            Assert.Equal(
-                Option.Some(input).SelectMany(MultiplyByTen),
-                MultiplyByTen(input));
-        }
-
-        [Theory]
         [MemberData(nameof(EqualOptions))]
         public void EqualsReturnsTrueForEqualOptions(Option<int> lhs, Option<int> rhs)
         {
