+++ Funcky/Compatibility/ExceptionUtilities.cs
