+++ Funcky/Monads/Either/Either.Core.cs
@@ -48,19 +48,6 @@ namespace Funcky.Monads
         public static Either<TLeft, TRight> Right(TRight right) => new(right);
 
         [Pure]
-        public Either<TLeft, TResult> Select<TResult>(Func<TRight, TResult> selector)
-            => Match(
-                Either<TLeft, TResult>.Left,
-                right => Either<TLeft, TResult>.Right(selector(right)));
-
-        [Pure]
-        public Either<TLeft, TResult> SelectMany<TEither, TResult>(Func<TRight, Either<TLeft, TEither>> eitherSelector, Func<TRight, TEither, TResult> resultSelector)
-            => Match(
-                Either<TLeft, TResult>.Left,
-                right => eitherSelector(right).Select(
-                    selectedRight => resultSelector(right, selectedRight)));
-
-        [Pure]
         public TMatchResult Match<TMatchResult>(Func<TLeft, TMatchResult> left, Func<TRight, TMatchResult> right)
             => _side switch
             {
