+++ Funcky/Monads/Lazy/Lazy.Monad.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.Contracts;
 
 namespace Funcky.Monads
 {
-    public static class LazyExtensions
+    public static partial class LazyExtensions
     {
         [Pure]
         public static Lazy<TResult> Select<T, TResult>(this Lazy<T> lazy, Func<T, TResult> selector)
