+++ Funcky/Monads/Result/Result.Core.cs
@@ -64,25 +64,6 @@ namespace Funcky.Monads
         }
 
         [Pure]
-        public Result<TResult> Select<TResult>(Func<TValidResult, TResult> selector)
-            => Match(
-                error: error => new Result<TResult>(error),
-                ok: value => Result.Ok(selector(value)));
-
-        [Pure]
-        public Result<TResult> SelectMany<TResult>(Func<TValidResult, Result<TResult>> resultSelector)
-            => Match(
-                error: error => new Result<TResult>(error),
-                ok: result => resultSelector(result));
-
-        [Pure]
-        public Result<TResult> SelectMany<TSelectedResult, TResult>(Func<TValidResult, Result<TSelectedResult>> selectedResultSelector, Func<TValidResult, TSelectedResult, TResult> resultSelector)
-            => Match(
-                error: error => new Result<TResult>(error),
-                ok: result => selectedResultSelector(result).Select(
-                    maybe => resultSelector(result, maybe)));
-
-        [Pure]
         public TMatchResult Match<TMatchResult>(Func<TValidResult, TMatchResult> ok, Func<Exception, TMatchResult> error)
             => _error is null
                 ? ok(_result)
